定数バッファ構造体と入力レイアウトを定義するヘッダーを作成、
シェーダーが構造体を持つことが出来るように修正
