+++ 3DAction_Project/3DAction_Project/ConstantBufferManager.cpp
@@ -1,4 +1,8 @@
-﻿// 必須ヘッダー
+﻿
+// ======================================
+// ヘッダー
+// ======================================
+// 必須ヘッダー
 #include "ConstantBufferManager.h" // 自分のヘッダー
 // DirectXヘッダー
 #include <d3d11.h>        // DirectXのAPIヘッダー
