+++ 3DAction_Project/3DAction_Project/ShaderStructs.h
@@ -1,0 +1,31 @@
+﻿#pragma once
+// =======================================
+// シェーダー関連で必要な構造体
+// =======================================
+
+
+// =======================================
+// ヘッダー
+// =======================================
+#include <string>
+
+
+// =======================================
+// 構造体
+// =======================================
+
+// 定数バッファの情報構造体
+struct ConstantBufferInfo {
+	std::string name = "";  // 定数バッファの名前
+	int registerNumber = 0; // レジスタ番号
+	size_t size = 0;        // 大きさ１６の倍数にする
+};
+
+// 入力レイアウト構造体
+struct InputLayoutInfo {
+	std::string semanticName = ""; // セマンティックの名前
+	int semanticIndex = 0;         // セマンティックの番号
+	int inputSlot = 0;             // スロット番号
+	int format = 0;                // データの形式 (DXGI)
+};
+
